Start animation immediately after pressing 'a' key
After not using the FPU emulation anymore, animation can now start
right away after pressing the 'a' key. FPU emulation was causing a
crash when a timer interrupt happened whilst working with floating
point numbers. With the fixed point arithmetic introduction (and
getting rid of any float / double types in the code along with it),
time alignment of the animation in between timer interrupts is not
necessary anymore. Additionally to that, fast mode (1s timer tick) is
also possible while animation is active.